apollo-ai-m6/pyproject.toml
[project]
name = "apollo-ai-m6"
version = "0.1.0"
description = ""
authors = [
    {name = "solideos",email = "urim.chae@solideos.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pymssql (>=2.3.7,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "networkx (==3.0)",
    "scikit-learn (==1.6)",
    "pandas (>=2.3.2,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
